Make wxBasicString as simple and safe to use as possible
Don't try to write a general purpose class, but provide just the methods
that we need in our code.

This fixes the bug added in 3f127dee3a565f6e0e5afa3d7d9a11693105f14b
which resulted in a crash because a literal string, not a BSTR, was
passed to wxBasicString ctor.