Really fix spurious asserts in wxGTK wxDataViewCtrl::EditItem()
This replaces the changes of 28dd56a15862fafe6f8e25ef6cf4ba4be5031d70
which, for some reason, used a global variable for storing whether the
selection function had been already set or not, when this clearly is a
per-control (or per-selection, but this seems one and the same) bit of
information.

Replace global ms_firstTime with a wxDataViewCtrlInternal field to avoid
asserts as soon as EditItem() is called on more than one wxDataViewCtrl.

Closes #17946.