Remove state assigning from wxProgressDialog::GetHandle()
This was added in fbd2f90f9b8b5cb73f68be1f81b17c896d4339aa without any
explanation and probably was a copy-and-paste typo as it just doesn't
make any sense to change the state of the dialog in this accessor
function (and if the state doesn't change, then this assignment is just
completely useless).

Remove the apparently unnecessary assignment and also an unnecessary
temporary variable.