Fix recently introduced crash in ClippingBoxTestCase
Commit 3d83acde3e4fff171c4487315875cd6dc645c240 did fix a memory leak in
this test case, but at the price of introducing a crash due to deleting
the same pointer twice.

The real fix would be to change the code here to avoid returning a
pointer which sometimes needs to be deleted and sometimes must not, but
for now just add a crude check to avoid crashing.