Exclude unnecessary files in the submodules from release archives
Cut down on release archives size, somewhat, by excluding third party
libraries that we're never going to use.

The exception is libtiff which references all its files in its configure
that we do run and it's probably not worth to bother with paring it down
if this means maintaining changes to configure.