Try to fix macOS build using cmake after xlocale changes
Commit 4d22b6378c4e5088b3c92edd5a24d8dad203c9e7 removed the inclusion of
xlocale.h because it is not (and never was) needed under Linux with
glibc, but it is still needed under macOS, so this (silently) disabled
wxXLocale support under Mac when using configure and broke the build
when using cmake.

Fix both problems by using xlocale.h only if it's available, both in
configure and in cmake.